[#128070485] Fixed Windows build
Changes:

* Added GSAGE_API definition to make plugins work.
* Building OGRE as a DLL, otherwise plugins get different singleton
context and everything explodes.
* Updated Sol2 library to the latest version. Old one causes DataProxy
to crash.
* Fixed Makefile steps to work in CMD console.
* Fixed entity deletion from Engine (unallocate then getId).
* Patched sole.hpp a bit to work with `winsock.h` instead of
`winsock2.h`, otherwise getting lots of conflicts.
* Fixed SDL2 window handle definition for OGRE.
* Enabled direct `LuaEventProxy` handler for callbacks, that are not
created in a coroutine.
* Patched jsoncpp: `Json::Exception` conflicts with `Ogre::Exception`
for some reason. Renamed to `JsonException`.
* Made luarocks packager work on Windows.

+++ PlugIns/OgrePlugin/src/OgreConverters.cpp
@@ -60,7 +60,7 @@ namespace Gsage {
     if(src.empty())
       return false;
     Ogre::ColourValue res;
-    unsigned int value;
+    unsigned long value;
     try
     {
       value = std::stoul(src, nullptr, 16);
